requirements.txt
pandas
numpy
ta
python-telegram-bot==13.15
